This page enables the user to ask a question on OMQ (Open Mind Questions).
It contains 3 textbox fields: Title, Description, and Tags.
1 button: Post Your Question

The user cannot proceed to ask a question if textboxes are left empty.
He will be alerted with a small message near the textbox he left empty and a highlight of 2 seconds on the border of the textbox to catch his attention.